clang-query: add JSON output
add the ability to produce a JSON dump of the AST in clang-query via

> enable output json

this will produce a JSON array, containing one JSON object per bound
match. If detailed-ast is enabled, each object will contain the AST dump
under the "ast" key. If the output option "print" is enabled, the
pretty-printed C++ code will be contained under the "pretty" key.